[move] writing 패키지 경로 변경 #22
 community 패키지 밖으로 이동
