+++ wingle/src/main/java/kr/co/wingle/writing/Writing.java
@@ -1,4 +1,4 @@
-package kr.co.wingle.community.writing;
+package kr.co.wingle.writing;
 
 import javax.persistence.Column;
 import javax.persistence.DiscriminatorColumn;
