+++ wingle/src/main/java/kr/co/wingle/writing/WritingDto.java
@@ -1,4 +1,4 @@
-package kr.co.wingle.community.writing;
+package kr.co.wingle.writing;
 
 import java.time.LocalDateTime;
